+++ projects/net.splitcells.website.server/CHANGELOG.md
@@ -27,6 +27,7 @@ The changelog format can be found [here](../../src/main/md/net/splitcells/networ
 * **2023-05-17: \#s86**
   1. Deprecate Renderer interface.
   2. Make sub renderer interfaces ProjectRenderer, ProjectsRenderer and ProjectRendererExtension more uniform.
+  3. Provide a way to query title meta data from ProjectsRenderer.
 * **2023-01-22**:
   1. [**\\#196** Create sane default config and behaviour regarding temporary data for, i.e. logging:](https://github.com/www-splitcells-net/net.splitcells.network/issues/196)
      1. Store logs etc. at `~./.local/state/net.splitcells.dem/src/main/[file type]/[program name]/[date time of program start]/` by default.